Add file input instead of text field
- The previous text field in the GUI toolbar was replaced by a
VFileInput object, which enables searching for the files locally. Thus,
a validation of the entered .dat path is not required anymore. Moreover,
the VFileInput object is currently designed to only search for .dat
files.
- The corresponding file objects do not contain the full local path
(not possible due to security reasons). Therefore, we now store the .dat
file name, its file lines, its size and the last modified data as state
variables to match the new file input structure. Nonetheless, the user
can see the full file path when clicking on the VFileInput object.
- Improved docstring documentation of several functions.