Fix scaling and add global COS
- Fixed scaling problems when switching between files. Specifically,
there is a new update_vtu_render_window function which performs manual
scaling procedures for vtk components which do not scale automatically.
- Added global coordinate system which also scales with the problem.
- Introduced CTXT context object of the SERVER. This is a
server-side-only state object. We use this for the non-serializable
objects which do not have to be shared with the client. This resolves the
errors we received regarding these objects.
- Improved docstring documentation of various functions.